pyproject.toml
[project]
name = "simple-ai-agent"
version = "0.1.0"
description = "A simple AI agent framework"
authors = [{name = "Sergei Konovalov", email = "l0kifs91@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.4",
    "loguru>=0.7.3",
    "openai>=2.1.0",
    "pydantic-settings>=2.11.0",
]
